Fixing settings
